+++ lib/index.d.ts
@@ -1,22 +1,14 @@
 export declare const Actions: (event: import("./types.js").EVENT, options: import("./types.js").Config) => Promise<{
     created: number;
     failed: number;
+    failed_msg: never[];
 }>;
 export declare const ActionsList: import("./types.js").Action[];
-export declare const Commands: (CommandEvent: import("./types.js").CommandEvent) => Promise<{
+export declare const Commands: (CommandEvent: import("./types.js").CommandEvent) => Promise<void | import("./types.js").Result | {
     qoyod_total: number;
     repzo_total: number;
     disabled: number;
     failed: number;
-} | {
-    created: number;
-    failed: number;
-} | {
-    qoyod_total_families: number;
-    repzo_total: number;
-    created_families: number;
-    created_secondary_units: number;
-    updated: number;
-    failed: number;
+    failed_msg: any[];
 }>;
 export declare const CommandsList: import("./types.js").Command[];