+++ lib/util.d.ts
@@ -1,3 +1,4 @@
+import Repzo from "repzo";
 interface Params {
     [key: string]: any;
 }
@@ -12,4 +13,7 @@ export declare const _fetch: (baseUrl: string, path: string, headers?: Headers |
 export declare const _create: (baseUrl: string, path: string, body: Data, headers?: Headers | undefined, params?: Params | undefined) => Promise<any>;
 export declare const _update: (baseUrl: string, path: string, body: Data, headers?: Headers | undefined, params?: Params | undefined) => Promise<any>;
 export declare const _delete: (baseUrl: string, path: string, headers?: Headers | undefined, params?: Params | undefined) => Promise<any>;
+export declare const update_bench_time: (repzo: Repzo, app_id: string, key: string, value: string) => Promise<void>;
+export declare const updateAt_query: (QUERY: string, options_formData: any, bench_time_key: string) => string;
+export declare const get_data_from_qoyod: (_path: string, default_res: any, serviceEndPoint: string, serviceApiKey: string, query?: string | undefined) => Promise<any>;
 export {};