+++ src/commands/index.ts
@@ -1,13 +1,13 @@
-import { Config, Command, CommandEvent } from "./../types";
+import { Config, Command, CommandEvent, Result } from "./../types";
 import { addClients, updatedInactiveClients } from "./client.js";
 import { addProducts } from "./product.js";
 import { sync_categories } from "./category.js";
 import { sync_measureunits } from "./measureunit.js";
 import { sync_inventory } from "./inventory.js";
 import { adjust_inventory } from "./adjust_inventory.js";
-import { sync_invoice } from "./invoice.js";
 import { sync_taxes } from "./tax.js";
 import { sync_measureunit_family } from "./measureunit_family.js";
+import { join } from "./join.js";
 import { EVENT } from "./../types";
 export const commands = async (CommandEvent: CommandEvent) => {
   switch (CommandEvent.command) {
@@ -29,14 +29,19 @@ export const commands = async (CommandEvent: CommandEvent) => {
       return await sync_inventory(CommandEvent);
     case "adjust_inventory":
       return await adjust_inventory(CommandEvent);
-    case "sync_invoice":
-      return await sync_invoice(CommandEvent);
+    case "join":
+      return await join(CommandEvent);
     default:
       throw `Route: ${CommandEvent.command} not found`;
   }
 };
 
 export const commandsList: Command[] = [
+  {
+    command: "join",
+    name: "Join",
+    description: "",
+  },
   {
     command: "add_client",
     name: "Sync Clients",
@@ -80,11 +85,6 @@ export const commandsList: Command[] = [
   {
     command: "adjust_inventory",
     name: "Adjust Inventory",
-    description: "",
-  },
-  {
-    command: "sync_invoice",
-    name: "Sync Invoice",
     description: "",
   },
 ];