+++ src/commands/measureunit.ts
@@ -1,7 +1,14 @@
 import Repzo from "repzo";
 import DataSet from "data-set-query";
-import { EVENT, Config, CommandEvent } from "../types";
-import { _fetch, _create, _update, _delete } from "../util.js";
+import { EVENT, Config, CommandEvent, Result } from "../types";
+import {
+  _fetch,
+  _create,
+  _update,
+  _delete,
+  update_bench_time,
+  updateAt_query,
+} from "../util.js";
 
 interface QoyodUnit {
   id: number;
@@ -14,21 +21,48 @@ export interface QoyodUnits {
 }
 
 export const sync_measureunits = async (commandEvent: CommandEvent) => {
+  const repzo = new Repzo(commandEvent.app.formData?.repzoApiKey, {
+    env: commandEvent.env,
+  });
+  const commandLog = new Repzo.CommandLog(
+    repzo,
+    commandEvent.app,
+    commandEvent.command,
+  );
   try {
     console.log("sync_measureunits");
+
+    const new_bench_time = new Date().toISOString();
+    const bench_time_key = "bench_time_measureunit";
+
+    await commandLog.load(commandEvent.sync_id);
+    await commandLog
+      .addDetail("Repzo Qoyod: Started Syncing Measure Unit")
+      .commit();
+
     const nameSpace = commandEvent.nameSpace.join("_");
-    const result = {
+    const result: Result = {
       qoyod_total: 0,
       repzo_total: 0,
       created: 0,
       updated: 0,
       failed: 0,
+      failed_msg: [],
     };
+
     const qoyod_units: QoyodUnits = await get_qoyod_units(
       commandEvent.app.available_app.app_settings.serviceEndPoint,
       commandEvent.app.formData.serviceApiKey,
+      updateAt_query("", commandEvent.app.options_formData, bench_time_key),
     );
     result.qoyod_total = qoyod_units?.product_unit_types?.length;
+    await commandLog
+      .addDetail(
+        `${qoyod_units?.product_unit_types?.length} taxes changed since ${
+          commandEvent.app.options_formData[bench_time_key] || "ever"
+        }`,
+      )
+      .commit();
 
     const db = new DataSet([], { autoIndex: false });
     db.createIndex({
@@ -41,10 +75,6 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
       (unit: QoyodUnit) => `${nameSpace}_${unit.id}_1.0`,
     ); // ??
 
-    const repzo = new Repzo(commandEvent.app.formData?.repzoApiKey, {
-      env: commandEvent.env,
-    });
-
     const repzo_base_unit = await repzo.measureunit.find({
       parent: "nil",
       factor: 1,
@@ -53,13 +83,23 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
       repzo_base_unit?.data.length == 1
         ? repzo_base_unit?.data[0]._id
         : undefined;
-    if (!repzo_base_unit_id) throw new Error(`Repzo Base Unit was not found`);
+    if (!repzo_base_unit_id) {
+      await commandLog
+        .setStatus("fail", `Repzo Base Unit was not found`)
+        .commit();
+      throw new Error(`Repzo Base Unit was not found`);
+    }
 
     const repzo_units = await repzo.measureunit.find({
       "integration_meta.id": unit_query,
-      // project:["_id", "name", "integration_meta", "disabled", "email", "phone", "tax_number"]
     });
     result.repzo_total = repzo_units?.data?.length;
+    await commandLog
+      .addDetail(
+        `${repzo_units?.data?.length} taxes in Repzo was matched the integration.id`,
+      )
+      .commit();
+
     for (let i = 0; i < qoyod_units.product_unit_types.length; i++) {
       const qoyod_unit: QoyodUnit = qoyod_units.product_unit_types[i];
       const repzo_unit = repzo_units.data.find(
@@ -87,7 +127,12 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
           const created_unit = await repzo.measureunit.create(body);
           result.created++;
         } catch (e: any) {
-          console.log("Create Measure Unit Failed >> ", e.response, body);
+          console.log("Create Measure Unit Failed >> ", e?.response, body);
+          result.failed_msg.push(
+            "Create Measure Unit Failed >> ",
+            e?.response,
+            body,
+          );
           result.failed++;
         }
       } else {
@@ -104,18 +149,32 @@ export const sync_measureunits = async (commandEvent: CommandEvent) => {
             body,
           );
           result.updated++;
-        } catch (e) {
-          console.log("Update Measure Unit Failed >> ", e, body);
+        } catch (e: any) {
+          console.log("Update Measure Unit Failed >> ", e?.response, body);
+          result.failed_msg.push(
+            "Update Measure Unit Failed >> ",
+            e?.response,
+            body,
+          );
           result.failed++;
         }
       }
     }
 
     console.log(result);
+
+    await update_bench_time(
+      repzo,
+      commandEvent.app._id,
+      bench_time_key,
+      new_bench_time,
+    );
+    await commandLog.setStatus("success").setBody(result).commit();
     return result;
   } catch (e: any) {
     //@ts-ignore
-    console.error(e.response.data);
+    console.error(e?.response?.data);
+    await commandLog.setStatus("fail", e).commit();
     throw e?.response;
   }
 };
@@ -133,13 +192,7 @@ export const get_qoyod_units = async (
     );
     return qoyod_units;
   } catch (e: any) {
-    // code instead of msg
-    if (
-      e.response.data ==
-      "We could not retrieve your product_unit_types, we found nothing."
-    )
-      return { product_unit_types: [] };
-
+    if (e.response.status == 404) return { product_unit_types: [] };
     throw e;
   }
 };