+++ src/commands/product.ts
@@ -1,7 +1,14 @@
 import Repzo from "repzo";
 import DataSet from "data-set-query";
-import { EVENT, Config, CommandEvent } from "../types";
-import { _fetch, _create, _update, _delete } from "../util.js";
+import { EVENT, Config, CommandEvent, Result } from "../types";
+import {
+  _fetch,
+  _create,
+  _update,
+  _delete,
+  update_bench_time,
+  updateAt_query,
+} from "../util.js";
 
 interface QoyodProduct {
   id: number;
@@ -43,21 +50,49 @@ interface QoyodProducts {
 }
 
 export const addProducts = async (commandEvent: CommandEvent) => {
+  const repzo = new Repzo(commandEvent.app.formData?.repzoApiKey, {
+    env: commandEvent.env,
+  });
+  const commandLog = new Repzo.CommandLog(
+    repzo,
+    commandEvent.app,
+    commandEvent.command,
+  );
   try {
     console.log("addProducts");
+
+    const new_bench_time = new Date().toISOString();
+    const bench_time_key = "bench_time_product";
+
+    await commandLog.load(commandEvent.sync_id);
+    await commandLog
+      .addDetail("Repzo Qoyod: Started Syncing Products")
+      .commit();
+
     const nameSpace = commandEvent.nameSpace.join("_");
-    const result = {
+    const result: Result = {
       qoyod_total: 0,
       repzo_total: 0,
       created: 0,
       updated: 0,
       failed: 0,
+      failed_msg: [],
     };
+
     const qoyod_products: QoyodProducts = await get_qoyod_products(
       commandEvent.app.available_app.app_settings.serviceEndPoint,
       commandEvent.app.formData.serviceApiKey,
+      updateAt_query("", commandEvent.app.options_formData, bench_time_key),
     );
     result.qoyod_total = qoyod_products?.products?.length;
+    await commandLog
+      .addDetail(
+        `${qoyod_products?.products?.length} products changed since ${
+          commandEvent.app.options_formData[bench_time_key] || "ever"
+        }`,
+      )
+      .commit();
+
     const db = new DataSet([], { autoIndex: false });
     db.createIndex({
       id: true,
@@ -85,23 +120,38 @@ export const addProducts = async (commandEvent: CommandEvent) => {
     const product_query = qoyod_products?.products.map(
       (product: QoyodProduct) => `${nameSpace}_${product.id}`,
     ); // ??
-    const repzo = new Repzo(commandEvent.app.formData?.repzoApiKey, {
-      env: commandEvent.env,
-    });
+
     const repzo_products = await repzo.product.find({
       "integration_meta.id": product_query,
       withVariants: true,
     });
     result.repzo_total = repzo_products?.data?.length;
+    await commandLog
+      .addDetail(
+        `${repzo_products?.data?.length} products in Repzo was matched the integration.id`,
+      )
+      .commit();
 
     const repzo_categories = await repzo.category.find({ per_page: 50000 });
+    await commandLog
+      .addDetail(`${repzo_categories?.data?.length} product categories`)
+      .commit();
     const repzo_measureunits = await repzo.measureunit.find({
       per_page: 50000,
     });
+    await commandLog
+      .addDetail(`${repzo_measureunits?.data?.length} measure units`)
+      .commit();
     const repzo_taxes = await repzo.tax.find({ per_page: 50000 });
+    await commandLog.addDetail(`${repzo_taxes?.data?.length} taxes`).commit();
     const repzo_measureunit_family = await repzo.measureunitFamily.find({
       per_page: 50000,
     });
+    await commandLog
+      .addDetail(
+        `${repzo_measureunit_family?.data?.length} measure unit families`,
+      )
+      .commit();
 
     for (let i = 0; i < qoyod_products.products.length; i++) {
       const qoyod_product: QoyodProduct = qoyod_products.products[i];
@@ -125,6 +175,9 @@ export const addProducts = async (commandEvent: CommandEvent) => {
         console.log(
           `Update product Failed >> Category with integration_meta.id: ${nameSpace}_${qoyod_product.category_id} was not found`,
         );
+        result.failed_msg.push(
+          `Update product Failed >> Category with integration_meta.id: ${nameSpace}_${qoyod_product.category_id} was not found`,
+        );
         result.failed++;
         continue;
       }
@@ -138,6 +191,9 @@ export const addProducts = async (commandEvent: CommandEvent) => {
         console.log(
           `Update product Failed >> MeasureUnit with integration_meta.id: ${nameSpace}_${qoyod_product.unit_type} was not found`,
         );
+        result.failed_msg.push(
+          `Update product Failed >> MeasureUnit with integration_meta.id: ${nameSpace}_${qoyod_product.unit_type} was not found`,
+        );
         result.failed++;
         continue;
       }
@@ -148,6 +204,9 @@ export const addProducts = async (commandEvent: CommandEvent) => {
       );
       if (!measureunit_family) {
         console.log(
+          `Update product Failed >> MeasureUnit Family with integration_meta.id: ${nameSpace}_${qoyod_product.sku} was not found`,
+        );
+        result.failed_msg.push(
           `Update product Failed >> MeasureUnit Family with integration_meta.id: ${nameSpace}_${qoyod_product.sku} was not found`,
         );
         result.failed++;
@@ -163,6 +222,13 @@ export const addProducts = async (commandEvent: CommandEvent) => {
       );
       if (!tax) {
         console.log(
+          `Update product Failed >> Tax with integration_meta.id: ${nameSpace}_${
+            qoyod_product.tax_id
+          }_${
+            qoyod_product.is_inclusive ? "inclusive" : "additive"
+          } was not found`,
+        );
+        result.failed_msg.push(
           `Update product Failed >> Tax with integration_meta.id: ${nameSpace}_${
             qoyod_product.tax_id
           }_${
@@ -227,7 +293,12 @@ export const addProducts = async (commandEvent: CommandEvent) => {
           const created_product = await repzo.product.create(body);
           result.created++;
         } catch (e: any) {
-          console.log("Create product Failed >> ", e.response, body);
+          console.log("Create product Failed >> ", e?.response, body);
+          result.failed_msg.push(
+            "Create product Failed >> ",
+            e?.response,
+            body,
+          );
           result.failed++;
         }
       } else {
@@ -258,17 +329,32 @@ export const addProducts = async (commandEvent: CommandEvent) => {
             body,
           );
           result.updated++;
-        } catch (e) {
+        } catch (e: any) {
           console.log("Update product Failed >> ", e, body);
+          result.failed_msg.push(
+            "Update product Failed >> ",
+            e?.response,
+            body,
+          );
           result.failed++;
         }
       }
     }
+
     console.log(result);
+
+    await update_bench_time(
+      repzo,
+      commandEvent.app._id,
+      bench_time_key,
+      new_bench_time,
+    );
+    await commandLog.setStatus("success").setBody(result).commit();
     return result;
   } catch (e: any) {
     //@ts-ignore
-    console.error(e.response.data);
+    console.error(e?.response?.data);
+    await commandLog.setStatus("fail", e).commit();
     throw e?.response;
   }
 };
@@ -286,13 +372,7 @@ export const get_qoyod_products = async (
     );
     return qoyod_products;
   } catch (e: any) {
-    if (
-      // code instead of msg
-      e.response.data ==
-      "We could not retrieve your products, we found nothing."
-    )
-      return { products: [] };
-
+    if (e.response.status == 404) return { products: [] };
     throw e;
   }
 };