+++ src/util.ts
@@ -1,4 +1,5 @@
 import axios from "axios";
+import Repzo from "repzo";
 
 interface Params {
   [key: string]: any;
@@ -19,10 +20,7 @@ export const _fetch = async (
   params?: Params,
 ) => {
   try {
-    const res = await axios.get(baseUrl + path, {
-      params,
-      headers,
-    });
+    const res = await axios.get(baseUrl + path, { params, headers });
     return res.data;
   } catch (e) {
     throw e;
@@ -78,6 +76,61 @@ export const _delete = async (
     });
     return res.data;
   } catch (e) {
+    throw e;
+  }
+};
+
+export const update_bench_time = async (
+  repzo: Repzo,
+  app_id: string,
+  key: string,
+  value: string,
+): Promise<void> => {
+  try {
+    const res = await repzo.integrationApp.update(app_id, {
+      // options_formData: { [key]: value },
+      [`options_formData.${key}`]: value,
+    });
+    console.log(res);
+  } catch (e) {
+    throw e;
+  }
+};
+
+export const updateAt_query = (
+  QUERY: string,
+  options_formData: any,
+  bench_time_key: string,
+): string => {
+  try {
+    QUERY = QUERY || "";
+    if (options_formData && options_formData[bench_time_key]) {
+      QUERY += `${QUERY ? "&" : ""}q[updated_at_gteq]=${
+        options_formData[bench_time_key]
+      }`;
+    }
+    return QUERY;
+  } catch (e) {
+    throw e;
+  }
+};
+
+export const get_data_from_qoyod = async (
+  _path: string,
+  default_res: any, // if no data was found
+  serviceEndPoint: string,
+  serviceApiKey: string,
+  query?: string,
+): Promise<any> => {
+  try {
+    const result: any = await _fetch(
+      serviceEndPoint,
+      `/${_path}${query ? query : ""}`,
+      { "API-KEY": serviceApiKey },
+    );
+    return result;
+  } catch (e: any) {
+    if (e.response.status == 404) return default_res;
     throw e;
   }
 };