+++ changelog.md
@@ -9,6 +9,12 @@
 - [./src/commands/measureunit_family.ts] new Command: Measure Unit Family @maramalshen
 - [./src/commands/product] handle product.track_quantity as product.frozen @maramalshen
 - [./src/actions/create_invoice] create invoice in Qoyod @maramalshen
+- [./src/actions/create_payment] create payment in Qoyod @maramalshen
+- Handle Bench time in the Commands @maramalshen
+- Command Logs @maramalshen
+- Action Logs @maramalshen
+- [./src/commands/join] new Command: Join @maramalshen
+- [./src/actions/create_invoice] read key invoice.status from settings @maramalshen
 
 ### Changed